Add the ability to (optionally) tag comments with a keyword.
Entire tags can then be hidden. The idea is that we use a `final' tag
for comments that we want to address for the final version. Once the
draft is done, we can turn on display of the `final' tag.

Any comments not tagged get the `default' tag.

In my to-do list, find out a way to turn on *all* tags.

If people like this, we can port it to the main repo.