Clarify caption.
The actual effect that can be observed is about more than just the initial peak:
it so happens that on the LHS plot a number of in-process iterations (~1-30) are
slower than what comes after, so are classified as a changepoint segment. The
previous text didn't draw the reader's eyes to this.