Fix figure scaling.
The problem is that the figure titles are different widths, but both are wider
than the main graph. This makes the scaling look different. That stumped me for
a minute. Then I loaded both PDFs into Inkscape, chose the width and height of
the biggest and applied that to the smallest. Voila -- fixed!